modified models and did some restructuring
