+++ up_orders_project/orders_app/models.py
@@ -1,49 +1,56 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 from django.db import models
 from django.contrib.auth.models import User
 
-# Create your models here.
-# Custom user
+# Models
+# Custom User Model
 class CustomUser(models.Model):
-    ROLE_CHOICES =[
-        ('Consumer', 'Consumer'),
-        ('Merchant', 'Merchant'),
+    ROLES = [
+        ("Merchant", "Merchant"),
+        ("Customer", "Customer"),
     ]
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    name = models.CharField(max_length=150, null=True)
-    role = models.CharField(max_length=10, choices=ROLE_CHOICES)
+    user = models.OneToOneField(User, on_delete=models.CASCADE, unique=True)
+    username = models.CharField(max_length=150, unique=True)
+    first_name = models.CharField(max_length=150)
+    last_name = models.CharField(max_length=150)
+    role = models.CharField(max_length=20, choices=ROLES)
 
-    def __str__(self):
-        return self.name
+    # def __str__(self):
+    #     return self.first_name + ' ' + self.last_name
     
-# Store belongs to a Merchant
+# Store
 class Store(models.Model):
     name = models.CharField(max_length=150)
-    address = models.CharField(max_length=150, null=True)
+    slug = models.SlugField(unique=True)
+    address = models.CharField(max_length=150)
+    city = models.CharField(max_length=100)
+    state = models.CharField(max_length=100)
+    zip = models.CharField(max_length=10)
     merchant = models.ForeignKey(CustomUser, on_delete=models.CASCADE)
 
     def __str__(self):
-        return self.name + ' ' + self.address
-
-# Item belongs to a store
+        return self.name + ', ' + self.address + ', ' + self.city
+    
+# Item
 class Item(models.Model):
-    CATEGORY_CHOICES = (
-        ('Starter', 'Starter'),
-        ('Beverage', 'Beverage'),
-        ('Main Course', 'Main Course'),
-        ('Dessert', 'Dessert'),
-    )
+    class Meta:
+        unique_together = ('name', 'store')
 
+    CATEGORIES = [
+        ("Starter", "Starter"),
+        ("Beverage", "Beverage"),
+        ("Cold Beverage", "Cold Beverage"),
+        ("Snack", "Snack"),
+        ("Main Course", "Main Course"),
+        ("Dessert", "Dessert"),
+        ("Side", "Side")
+    ]
     name = models.CharField(max_length=150)
-    category = models.CharField(
-        max_length=50,
-        choices=CATEGORY_CHOICES
-    )
-    price = models.DecimalField(max_digits=6, decimal_places=2, default=0)
+    description = models.TextField()
+    category = models.CharField(max_length=150, choices=CATEGORIES)
+    price = models.DecimalField(max_digits=5, decimal_places=2)
+    merchant = models.ForeignKey(CustomUser, on_delete=models.CASCADE)
     store = models.ForeignKey(Store, on_delete=models.CASCADE)
 
     def __str__(self):
-        return self.name
+        return self.category + ': ' + self.name
