--- up_orders_project/orders_app/signals.py
@@ -1,5 +1,0 @@
-from django.contrib.auth.models import User
-from django.db import models
-from tastypie.models import create_api_key
-
-models.signals.post_save.connect(create_api_key, sender=User)
